requirements.txt
# ============================================================================
# WCC AI Learning Series - Requirements
# ============================================================================
#
# Session 1 (AI Fundamentals & LLM APIs) - REQUIRED
# ============================================================================

google-cloud-aiplatform>=1.26.0
vertexai>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
requests>=2.31.0

# ============================================================================
# Optional: Alternative Platforms (Uncomment to use)
# ============================================================================

# OpenAI API
# openai>=1.0.0

# AWS Bedrock
# boto3>=1.40.0
# botocore>=1.43.0

# Anthropic Claude
# anthropic>=0.7.0

# Cohere
# cohere>=4.0.0

# ============================================================================
# Optional: Advanced Features (Uncomment as needed)
# ============================================================================

# Data privacy & security
# google-cloud-dlp>=3.33.0
# presidio-analyzer>=2.2.359
# presidio-anonymizer>=2.2.359

# Token counting
# tiktoken>=0.12.0

# ============================================================================
# Future Sessions (Will be added as we progress)
# ============================================================================

# Session 2: Prompt Engineering & Security
# Session 3: Introduction to RAG
# Session 4: AI Agents - Part 1
# Session 5: AI Agents - Part 2
# Session 6: Evaluation & Monitoring
